Remove dev/lib_metadata_checks.py (close #380)
1. Move main functionality into lib_ROI_overlaps.py
2. Update test_unit_parse_yokogawa_metadata.py
3. Move the plotting part into examples/tools/show_FOV_ROIs
