+++ examples/tools/show_FOV_ROIs/MeasurementDetail_2x2_well.mrf
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<bts:MeasurementDetail bts:Version="1.0" bts:OperatorName="Joel" bts:Title="20200812-CardiomyocyteDifferentiation14-Cycle1" bts:Application="" bts:BeginTime="2020-08-12T17:27:17.603876+02:00" bts:EndTime="2020-08-13T03:20:17.1789113+02:00" bts:MeasurementSettingFileName="20200812-Joel-CardiomyocyteDifferentiation14-Cycle1.mes" bts:ColumnCount="12" bts:RowCount="8" bts:TimePointCount="5" bts:FieldCount="72" bts:ZCount="42" bts:TargetSystem="CV7000 FD" bts:ReleaseNumber="R1.17.05" xmlns:bts="http://www.yokogawa.co.jp/BTS/BTSSchema/1.0">
+  <bts:MeasurementSamplePlate bts:Name="20200812-CardiomyocyteDifferentiation14-Cycle1" bts:WellPlateFileName="20200812-CardiomyocyteDifferentiation14-Cycle1.wpi" bts:WellPlateProductFileName="1009602002_Greiner_#655090.wpp" />
+  <bts:MeasurementChannel bts:Ch="1" bts:HorizontalPixelDimension="0.1625" bts:VerticalPixelDimension="0.1625" bts:CameraNumber="3" bts:InputBitDepth="16" bts:InputLevel="1000" bts:HorizontalPixels="2560" bts:VerticalPixels="2160" bts:FilterWheelPosition="6" bts:FilterPosition="1" bts:ShadingCorrectionSource="SC_BP445-45_40x_M10_CH01.tif" />
+  <bts:MeasurementChannel bts:Ch="2" bts:HorizontalPixelDimension="0.1625" bts:VerticalPixelDimension="0.1625" bts:CameraNumber="1" bts:InputBitDepth="16" bts:InputLevel="1000" bts:HorizontalPixels="2560" bts:VerticalPixels="2160" bts:FilterWheelPosition="4" bts:FilterPosition="2" bts:ShadingCorrectionSource="SC_BP675-30_40x_M10_CH02.tif" />
+  <bts:MeasurementChannel bts:Ch="3" bts:HorizontalPixelDimension="0.1625" bts:VerticalPixelDimension="0.1625" bts:CameraNumber="2" bts:InputBitDepth="16" bts:InputLevel="2000" bts:HorizontalPixels="2560" bts:VerticalPixels="2160" bts:FilterWheelPosition="5" bts:FilterPosition="1" bts:ShadingCorrectionSource="SC_BP525-50_40x_M10_CH03.tif" />
+</bts:MeasurementDetail>
