+++ examples/tools/show_FOV_ROIs/README.md
@@ -1,0 +1,1 @@
+NOTE: running this script requires `matplotlib`
